Fix a bug getting tuple number in stats of tuple-merge. Also turn-off some debug info in "articles".

+++ src/ee/anticache/AntiCacheDB.h
@@ -67,7 +67,7 @@ class AntiCacheBlock {
             return (m_payload.tableName);
         }
 
-        inline int32_t getSize() const {
+        inline long getSize() const {
             return m_size;
         }
         inline char* getData() const {
